requirements.txt
requests
Pillow
colorgram.py